I hope this works

+++ FiveNightsAtGorillas/Managers/DoorManager.cs
@@ -169,14 +169,14 @@ namespace FiveNightsAtGorillas.Managers.DoorAndLight
                 {
                     object[] value = new object[] { PhotonData.Close, 0.772f };
                     RaiseEventOptions options = new RaiseEventOptions() { CachingOption = EventCaching.DoNotCache, Receivers = ReceiverGroup.All };
-                    PhotonNetwork.RaiseEvent((byte)PhotonData.RightDoor, value, options, SendOptions.SendReliable);
+                    PhotonNetwork.RaiseEvent(PhotonData.RightDoor, value, options, SendOptions.SendReliable);
                     return;
                 }
                 else
                 {
                     object[] value = new object[] { PhotonData.Open, 2.272f };
                     RaiseEventOptions options = new RaiseEventOptions() { CachingOption = EventCaching.DoNotCache, Receivers = ReceiverGroup.All };
-                    PhotonNetwork.RaiseEvent((byte)PhotonData.RightDoor, value, options, SendOptions.SendReliable);
+                    PhotonNetwork.RaiseEvent(PhotonData.RightDoor, value, options, SendOptions.SendReliable);
                     return;
                 }
             }
@@ -186,14 +186,14 @@ namespace FiveNightsAtGorillas.Managers.DoorAndLight
                 {
                     object[] value = new object[] { PhotonData.Close, 0.75f };
                     RaiseEventOptions options = new RaiseEventOptions() { CachingOption = EventCaching.DoNotCache, Receivers = ReceiverGroup.All };
-                    PhotonNetwork.RaiseEvent((byte)PhotonData.LeftDoor, value, options, SendOptions.SendReliable);
+                    PhotonNetwork.RaiseEvent(PhotonData.LeftDoor, value, options, SendOptions.SendReliable);
                     return;
                 }
                 else
                 {
                     object[] value = new object[] { PhotonData.Open, 2.35f };
                     RaiseEventOptions options = new RaiseEventOptions() { CachingOption = EventCaching.DoNotCache, Receivers = ReceiverGroup.All };
-                    PhotonNetwork.RaiseEvent((byte)PhotonData.LeftDoor, value, options, SendOptions.SendReliable);
+                    PhotonNetwork.RaiseEvent(PhotonData.LeftDoor, value, options, SendOptions.SendReliable);
                     return;
                 }
             }
